test: split common container tests
